Move encodeUint256 & decodeUint256 out of assetDataUtils since we don't want them exported
